+++ addon/controllers/index.js
@@ -1,5 +1,5 @@
 import Controller from "@ember/controller";
-import { inject as service } from "@ember/service";
+import { service } from "@ember/service";
 import { tracked } from "@glimmer/tracking";
 
 export default class IndexController extends Controller {
@@ -27,10 +27,10 @@ export default class IndexController extends Controller {
 
   get documentFilters() {
     let filters = {
-      category: this.category,
+      categories: this.category,
       tags: this.tags.length ? this.tags.join(",") : undefined,
       marks: this.marks.length ? this.marks.join(",") : undefined,
-      search: this.search,
+      query: this.search,
       activeGroup: this.activeGroup,
     };