+++ addon/utils/error-handler.js
@@ -1,5 +1,5 @@
 import { getOwner, setOwner } from "@ember/application";
-import { inject as service } from "@ember/service";
+import { service } from "@ember/service";
 
 export class ErrorHandler {
   @service notification;