+++ tests/dummy/config/icons.js
@@ -1,5 +1,11 @@
 module.exports = function () {
   return {
-    "free-solid-svg-icons": ["stamp", "bullhorn", "heart", "dollar-sign"],
+    "free-solid-svg-icons": [
+      "stamp",
+      "bullhorn",
+      "heart",
+      "dollar-sign",
+      "file-code",
+    ],
   };
 };